Fix indentation in DB load code.
As written, we only load the state of one arbitrary user and store it in Persons. This usually doesn't matter, but after creating a fresh VM, we overwrite the entire WCA datastore since we don't have the previous copy of the WCA DB. So this wiped out all of the state rankings, by not copying state from User to Person (and then to RankSingle / RankAverage).